requirement.txt
Django
djangorestframework
python-dotenv
PyPDF2
requests
django-cors-headers
